Optimizations in the contracts Co-authored-by: Giorgio Bernardi <giorgiobernardi@users.noreply.github.com>

+++ contracts/libraries/DomainLib.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.28;
 library DomainLib {
     struct DomainConfig {
         string domain;
-        uint256 powerLevel;
+        uint128 powerLevel;
         uint256 expiryDate;
     }
 }
